Introduce scopes in the compiler (#28)
Add the concept of "scopes" into the `pavex` compiler.
As it stands, our scope tree is at most two levels deep—the root and the
leaves, one for each registered request handler.

I performed a bit of a refactoring when it comes to user components. We
folded all the different moving parts into a single module and hid that
complexity from the rest of the compiler by having a single `pub(crate)`
`UserComponentDb`.
The `RawCallableIdentifiersDb` is gone, while `ResolvedPathDb` became an
internal implementation detail.