Detect at compile-time if T in RouteParams<T> tries to extract a non-existing route paremeter (#33)
We introduce a `RouteParams` attribute macro which derives
`serde::Deserialize`, `serde::Serialize` and `StructuralDeserialize`.
The latter is a new marker trait that we introduce to tag types that use
the "out-of-the-box" deserialization strategy provided by `serde`. For
those types we perform additional static analysis: we ensure that all
their fields are named after a route parameter when they are used in
combination with the `RouteParams` extractor.