Borrow checking — Part 0 (#38)
Up until now, `pavex` did not take into account the borrow checker when
determining if a blueprint was valid and/or generating code. As a
result, it would gladly accept blueprints that had no chance of
generating code that compiled.
This PR is a first step in rectifying this shortcoming: we introduce a
new analysis pass on `CallGraph`s that tries to determine if there is a
way to generate code that passes the borrow checker.

We add a new knob for users, `CloningStrategy`, which can be used to
specify if the type returned by a constructor can or cannot be cloned by
the framework.
Based on this configuration, the analysis can introduce appropriate
calls to `.clone()` to transform the `CallGraph` into something that
will generate valid code.

This first draft is **extremely** rudimental and the code is a bit of a
mess. It does not take into account the branching introduced by `match`
statements and it does not yet detect basic failures (e.g. trying to
consume the same type multiple times).
But it's a start! A good enough start to highlight that our current
representation of borrowing in the `CallGraph` is extremely
annoying—therefore I'll merge this in and then get started on a
refactoring that moves the "borrows" information from a dedicated node
to metadata on the edge connected the borrowed and the borrower.