Borrow checking—Part 2 (#41)
Enhance the borrow checking passes to take into account control flow—if
a type is consumed by two dependents, are those two dependents ever
going to be executed one after the other?

E.g. one of those dependents is called in the `Ok` branch of a `match`
statement, while the other is triggered on the `Err` branch. All good!

Control-flow awareness extends to our error reporting as well—when there
is a conflict, we take care to only report the set of consumers that are
actually competing to consume a certain type within the same control
flow execution path.