Borrow checking—Part 4: polishing (#45)
This is a polishing PR: we add several more tests to verify how the
borrow checking logic behaves when the node under contention is an input
parameter (with the respect to the call graph being examined) or a type
provided by the framework.

As it's often the case with polishing, we ended up exposing some of the
short-comings of our current modeling: built-in framework types (e.g.
`Request<Body>` or `RawRouteParams`) were not properly modeled and, as a
consequence, the error message could not "detect" that it was dealing
with one of those.
In particular, we don't want to ask users to allow cloning of a built-in
framework type—they can't do anything about their clonability, it's a
bad suggestion.

We refactored `CallGraphNode` to capture the expected source of all
input parameters (either application state (singletons), framework
built-in types or user-provided) and we can now use it when crafting
errors.

By trying to clone singletons we uncovered another dormant issue: the
application state scope was a child of each request handler scopes, de
facto making it impossible to register transformers that would only
apply to a specific request handler call graph. We remediated the issue
by changing the shape of the graph (with some laborious ASCII art
edits).

Last but not least, we added a test that points at a short-coming that
we haven't yet resolved: `pavex` does not yet know how to reason about
types that contain references to other components.
It's definitely a solvable problem, but not one that we are going to
tackle right now.