Crate name resolution—Advanced mode (#48)
While working on
https://github.com/TechEmpower/FrameworkBenchmarks/pull/8162 I noticed
that `pavex` could not resolve local (i.e. `crate::XXX`) items
registered in a crate that was NOT the one rendering the `Blueprint`.

It all boils down to the difference between `std::env::var` and `env!`:
the former is evaluated when the code is run, the latter when it is
compiled.
In order to correctly resolve `crate::` items, we need to use `env!` and
make sure it's invoked from the code that is registering the local
item—i.e. it has to be part of the desugaring of the `f!` macro.

This PR adds a bunch more tests to make sure that we cover all bases. In
particular, it uncovered a limitation of `rustdoc`'s JSON format: there
are situations where we can't disambiguate between two different
versions of the same crate.
I've opened an MCP in the compiler repository to drive a change that
should resolve the issue. In the meantime, the test is marked as
"ignore: true".