Instrumentation, Chrome output and performance tuning (#54)
This PR is mostly the result of a few "instrument, measure, optimise"
cycles, looking to understand where `pavex_cli` spends its time when
generating application code.

As it turns out, with a warm cache, the vast majority of the time is
spent computing the JSON docs for the local crates. As it stands, this
is not something that we can optimise on our own.
We have two options, which both require upstream work:

1. Make the JSON docs generation faster;
2. Compute multiple JSON docs in parallel (up to the number of threads
available on the machine).

I don't have any concrete leads on 1., while 2. depends on adding
`--output-format` support to `cargo doc`. See
https://github.com/rust-lang/cargo/issues/12103

While hunting inefficiencies down, I added instrumentation around the
key steps of the compiler as well as enriched the `--debug` option with
a disk-based output that can be loaded in the Chrome tracing viewer
(chrome://tracing or ui.perfetto.dev).

You get this kind of visualisation:

<img width="1242" alt="immagine"
src="https://github.com/LukeMathWalker/pavex/assets/20745048/bd5381c3-d1db-48ce-9b59-eadb956ff4a3">

This is an execution trace with a warm cache. You can clearly see how
the lion share of the execution time it spend in `_compute_crate_docs`.

Miscellaneous changes:
- Re-added the `realworld` scaffolding in the `examples` folder;
- The package graph and the SQLite connection are now created in
parallel. We save ~10ms on the overall execution time.