Concurrency: compute multiple JSON docs in parallel using `cargo doc` (#62)
Little revelation on
[Zulip](https://rust-lang.zulipchat.com/#narrow/stream/266220-rustdoc/topic/JSON.20metadata.20speed/near/359024243)
last week: we can already leverage `cargo` to compute docs in parallel
by using `RUSTDOCFLAGS` to influence the behaviour of `cargo doc`.

This PR changes our invocation to use `cargo doc` instead of `cargo
rustdoc`, therefore accepting an array of package ids as input rather
than a single one.
We then expose this capability to the rest of the codebase by
integrating it with our two caching layers (in-memory and on disk via
SQLite).
We finally leverage it in our `UserComponentDb`. The previous behaviour
would compute the docs for a crate on the fly, as we found out we needed
them. That's still necessary, but we can perform a cheap scan of all the
paths that have been registered to try to pre-compute the relevant docs
ahead of time, therefore amortizing the cost.


Miscellaneous changes:
- Adopt the strategy from #61 for out test suite as well
- Each test should use only one core, otherwise we end up using more
capacity than the one we have on the host machine