QueryParams extractor (#65)
This PR adds a `QueryParams` extractor, built on top of
`serde_html_form`.
I've chosen `serde_html_form` instead of `serde_urlencoded` to get basic
support for sequences, an extremely common ask when working with query
parameters. The documentation covers the sequence syntax, including a
pointer to `serde_qs` for more advanced usecases.

Another big shift is moving away from the `http::Request` type: it is no
longer injected by default from the framework. You can either inject the
body (currently naked, to be fixed in a follow-up PR) or `RequestHead`.
This ensures that constructors "lock" the minimum required set of
information when running, which is going to reduce the chance of borrow
checking errors on the body as well as allow us higher levels of
concurrency down the line.

Miscellaneous changes:

- Use Pavex as the project name spelling everywhere, consistently
- Update `skeleton` example