Use a pool of target directories for UI tests (#69)
Each UI test is its own `cargo` project with its own target directory.
This has now become a problem since there are ~70 UI tests and my laptop
is running out of space.

We introduce a pool of target directories and assign them dynamically to
each test case when they are run. This ensures that we have an upper
limit on the disk usage without compromising on test parallelism.