Add all articles missing endpoints. Fix a bug with local re-exports (#73)
We now model all the endpoints under the `/articles` prefix in the
Realworld example.

I've also re-organised the Realworld example to go for the "1 file 1
route handler" model, which makes things a bit more obvious in terms of
code organisation.
When doing so, I unveiled a bug in our handling of local glob re-exports
(i.e. `use internal_module::*`): we were adding the name of the source
module to the path of the discovered items.
The bug has been fixed, there is a regression test and now everything
works as expected.