[Realworld] Implementation of user-related routes (#79)
It's time to actually flesh out the logic inside the routes for the
Realworld example!
This PR covers the sign-up and the login endpoint. The password logic is
mostly in place, we have to expand on input validation in a follow-up
PR.
This PR includes as well a basic integration test setup, which I'll
expand upon going forward.

Other refactoring:

- The configuration structs have been moved into the `conduit_core`
crate, while the machinery to _load_ configuration stays in the
`api_server` crate. This allows us to attach constructors to the config
structs and register them against the blueprint.

On the way, I found (and fixed) a bunch of bugs:

- Our re-export logic only worked for re-exported modules, but it didn't
work for re-exported items (found out by trying to register a
constructor for `PgPool`)
- Our logic for method detection didn't take into account type aliases
- Deserializing JSON docs for "pathological" crates (e.g. `typenum`)
results in a recursion limit error

And a little feature: we now implement the `Error` trait for the error
variant returned by the `build_application_state` function in the
code-generated server SDK.