Fix date for vision update in README.md (#82)
*Description of changes:*
It's a very minor change but it confused me momentarily.

Feel free to decline this PR if you're a time traveler and linear time
is actually a myth 🤭


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.