Wrapping middleware #3: A working MVP (#86)
We now have a working MVP for wrapping middleware 🎉 

In particular:

- For each middleware, we generate an ad-hoc type that encapsulates the
long-lived state needed by the remaining part of the request processing
pipeline (i.e. the concrete type for `Next<T>`).
- The code generator has been updated accordingly and we now generate a
separate module for each request handler, encapsulating its middlewares,
their states and the final request handler.
- Further validation has been added to ensure that the `Next<_>`
parameter is used correctly in user-provided middlewares.

A few notes:

- We are now using `IntoFuture` for `Next` rather than `Future`. It
makes working with the actual `Future`-implementing type a bit more
cumbersome (you need to call `.into_future`) but the implementation is
simpler. Something to think about for the future.
- This PR doesn't try to stress our borrow checking logic against the
new complexity that we are trying to model. We might have to extend our
capabilities to guarantee useful errors in most circumstances.
- Test coverage has to be significantly expanded.