The state of a middleware can capture references and components with lifetime parameters (#90)
One of my goals with Pavex is reducing/removing the friction of having
to clone or introduce synchronization when working with pieces of
request-scoped or transient state: any type that won't outlive a single
request and can be logically classified as "task-local" state for the
request handling task.

That was "easy" enough when we only had handlers—we only had to verify
that the borrow-checker would be happy with the generated code.
The introduction of wrapping middlewares made things trickier. A
request-scoped type might be constructed at the beginning of the request
lifecycle and then needed again in the final handler: we need to trickle
it down through each middleware invocation, which in turn implies that
we need to capture it in the code-generated "state" type we associate to
each middleware.
This worked well enough for owned types, but it broke down badly if
there were references or lifetime parameters—e.g. if you wanted to touch
the query parameters in multiple points of the chain without allocating
more than strictly necessary.

This PR fixes the issue: our handling of lifetimes throughout the
compiler had to become a lot more sophisticated, but we can now generate
the expected lifetime parameters on the middleware state and avoid
unnecessary clones.

Bonus points: I removed a layer of unnecessary boxing when using
middlewares.
The only auto-boxing done by Pavex, now, is the one needed to map to a
single response type.