Realworld: logging middleware (#92)
Now that the middleware infrastructure is in place, I can resume working
on the Realworld implementation. In particular, on the logging
middleware.

This PR fleshes out the primary fields, following OpenTelemetry's naming
convention.
It adds a new primitive extractor, `MatchedRouteTemplate`, in order to
fulfil the requirements.

A few things are still missing though:

- Connection-level information, which is currently unavailable since we
aren't propagating it down from the `Server`;
- Middlewares do not run if an incoming request fails to match one of
the handlers that have been registered against the `Blueprint`. In order
words, you get no logs for `404 Not Found` requests or for `405 Method
Not Allowed`.

I'll try to lift both limitations in a follow-up PR.