Documentation intro (#104)
I'm figuring things out as I go along, but I think I have a good
structure in mind for the documentation: this PR adds the introduction
as well as the beginning of the quickstart tutorial.

The tutorial/guide/concept/api reference structure comes from
[Diataxis](https://diataxis.fr/), but I have not yet decided if I want
to follow it as a top-level structure (i.e. the state of this PR) or as
a nested structure (i.e. organise by topic and each topic then haves
various types of documentation).
I'll think about it once I get past the introductory tutorial.