Patch-based workflow for tutorials in Pavex's documentation (#112)
Snippets and error messages in tutorials tend to go stale: the codebase
changes and there is no automated mechanism to remind you which snippets
need to be updated.
This PR migrates Pavex's tutorial to a patch-based workflow: no snippet
is "hard-coded" in the docs. Everything is generated from a starter
project and a series of `git` patches, applied one after the other.
Each code example from the tutorial is "just" an include from one of the
materialized stages of the tutorial project. The same applies to error
messages: we auto-generate them by running commands against a specific
stage of the tutorial project.

To get this working properly, I had to make a few more changes:
- `pavex_cli` will now consider the `color` setting (and the shell
properties) when emitting the `ERROR` header. It used to be always
coloured, no matter what.
- We no longer pull the template for `pavex new` down from GitHub: it's
now embedded into the binary. This solves some annoying problems with
GitHub Actions (i.e. the commit SHA that an action checks out is a
detached state which doesn't exist on the corresponding branch,
resulting in errors). It also allows `pavex new` to work offline and
removes the need for a patched `cargo-generate`.