Add wrapping behaviour to code labels (#116)
Our error reporter had no wrapping behaviour for code labels: they would
happily go as far right as they wanted, often overflowing the assumed
width of the terminal they are being printed on.

This PR changes that behaviour in two ways:
- Labels are no longer forced to be right-positioned to the code span
they refer to. If it helps to keep them on one line, we now position
them to the left or centred (with this order of preference).
- If they would necessarily overflow the maximum allowed width, we
perform text wrapping.

Getting this right was... tricky. But we have a fairly extensive
collection of error messages in our UI tests and the new version seems
to work as expected.