chore: ignore lints on generated crate (#127)
*Description of changes:* When playing around with Pavex, I added this
to declutter my `rust-analyzer` suggestions without touching the
generated code. Thought it'd be interesting to propose upstream, since
it gives the user more options on what to avoid (vs changing a file that
will get overwritten). Another option would be to add these to the top
of the generated file along with `unused_imports`.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Luca Palmieri <20745048+LukeMathWalker@users.noreply.github.com>