Introduce ByteUnit (#147)
Manipulating bytes as `usize`s is... clunky. 
I find myself commenting on relevant lines with the value converted in
the most relevant unit (e.g. MB or GB). That's a clear-cut code smell if
I've ever seen one.

We introduce `ByteUnit` by re-exporting from the `ubyte` crate. It is
used in
[`rocket`](https://api.rocket.rs/master/rocket/data/struct.ByteUnit.html)
and it fits our needs perfectly.

⚠️ Breaking change: I renamed `max_n_bytes` to `max_size` in
`BodySizeLimit`. The type has also changed from `usize` to `ByteUnit`.
This impacts you if you were registering a custom constructor for
`BodySizeLimit`. The same applies to the `SizeLimitExceeded` error.