Add white logo for dark mode (#153)
*Description of changes:*
This adds a white version of the logo to `README.md` that will be
rendered when GitHub is in dark mode. This approach is documented in
[GitHub's
blog](https://github.blog/changelog/2021-11-24-specify-theme-context-for-images-in-markdown/).

<div align="center">
<img width="600px" alt="Screen Shot 2024-01-08 at 20 00 26"
src="https://github.com/LukeMathWalker/pavex/assets/10367662/01899d94-5313-43e5-b6e2-eb3b23a66917">
</div>

I created this version by using the original as a mask on a white layer.
This ensures pixels are only white and translucent.

---

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.