Middlewares: execution order + docs (#156)
While documenting the execution order of middlewares, I realised that
the existing implementation resulted in some unintuitive behaviour (i.e.
a middleware would apply to all routes in the same `Blueprint`, even if
the middleware was registered after a route).
This PR changes the behaviour to align with the user's intuition.

## Internals

Refactorings lead to other refactorings: I had to change the internal
representation of a `Blueprint` to fix the issue with middlewares. I
took the opportunity to remove the `internals` module from `pavex`
entirely. It is now in a separate crate, `pavex_bp_schema`, which is not
exposed through `pavex`'s public API. It is an (implicit) dependency of
`Blueprint::persist` and `Blueprint::load`.