Activation key (#173)
All beta testers are currently using Pavex via `git` dependencies,
either targeting a branch or a tag.
That's... not ideal, since `cargo` doesn't really understand semver when
using `git` dependencies.

To improve the experience, I want to start releasing Pavex to crates.io.
To restrict access, this PR introduces an activation key. You'll only be
prompted for this once; Pavex will remember it as you update the CLI
too.

The current verification mechanism is a stop-gap solution to rush out
crates.io releases. The final version will include a server with
per-user keys, tied to a subscription. But since we don't want to charge
closed beta testers, this is fine for now.