ApiKit (#179)
`pavex new` currently adds a `register_common_constructors` function to
the generated project. In there, we call `register` for all common
first-party constructors from `pavex`.

There's a gotcha: once a project has been scaffolded, we can't update
it. All updates to the template will only apply to new projects.

We introduce **kits** to solve this issue and ensure that users keep
benefiting from new first-party constructors.
Kits are bundles of constructors: the template will install `ApiKit`
and, when new versions of Pavex are released, you'll automatically get
updates to the set of defaults (unless you choose to opt out).

## Breaking changes

Kits needs a way to tune and configure constructors _before_ they're
registered with a `Blueprint`.
To support the requirement, we renamed all existing
`Route`/`Constructor`/`Fallback`/`WrappingMiddleware` to `Registered*`.
We then introduce a `Constructor` type which can be used to assemble a
constructor _without_ registering it immediately.