Cookies integration (#234)
We re-export `biscotti` as our cookie API, with one key difference: we
don't re-export `ResponseCookies` as is.
We instead expose our own `ResponseCookies` type that wraps around
`biscotti`'s one, constraining its generic lifetime to be `'static`
since that's what the vast majority of applications will need.

The integration is not yet ready for prime time—we need a section in the
guide. But that'll come in a follow-up PR.

Other changes:

- Pavex gained a top-level `error` module, where I added an
`UnexpectedError` type. Almost any application defines one as a
"catch-all" error variant, therefore it looked like a win to make it
first-party since we need it anyway for our own error enums.