testing: Use `cargo-mutants` to surface gaps in `pavex_session`'s test suite
Mutation testing highlighted a few gaps in the test suite for `pavex_session`
that had not been caught by looking at coverage on its own.
New tests have been added to catch all discovered "MISSED" mutants,
as well as guidance on how to rerun these tools when touching the crate
in the future.