breaking(pavex_session): Re-organize `pavex_session`'s public API to minimise friction for common operations
The previous API gave equal standing to client-side and server-side state manipulation,
while the latter is to be preferred and likely to be more widely used.
This change moves all server-related method to appear directly on the `Session`
type, while client-side method remain on dedicated accessors.

Capabilities are unchanged, but the user-experience should be more pleasant.