fix: Always check if pavexc dependencies are installed
The check used to be run on installation, but this causes issues
with CI caching: if the binary is restored from a cache in a new runner,
the environment hasn't gone through the setup but the flow won't
trigger since the binary is already where we expect it to be.

We now run the setup flow unconditionally.