Introduce attribute-based components (#478)
We introduce new attribute macros to mark methods or functions as
constructors:

- `#[pavex::constructor]`
- `#[pavex::request_scoped]`
- `#[pavex::transient]`
- `#[pavex::singleton]`

We also introduce a new `Blueprint` method, `import`, to determine which
annotated constructors are in scope at any given time, mimicking the
behaviour of Rust's `use` (as well as its syntax).

## Internal refactors

### Diagnostics

Beyond the changes required to find, validate and consider the new
components, I also had to extensively rework our diagnostic system to
provide a higher-level API, capable of abstracting over the way a
component was registered.
The outcome is a net positive change: working with spans is a lot easier
now, and the code has clearer intent than before.

### `OrderedCallGraph`

In order to simplify debugging, `OrderedCallGraph` has been reworked to
embed the ordering the in the node indices rather than using an external
`id <> position` mapping. This caused a lot of churn in our diagnostic
snapshots.